// packages/fhir-r4/src/v4.0.1/valuesetfilter/types.ts
import type { Extension, Element } from "../core/types";

/** Generated from FHIR JSON Schema */

/** A ValueSet resource instance specifies a set of codes drawn from one or more code systems, intended for use in a particular context. Value sets link between [[[CodeSystem]]] definitions and their use in [coded elements](terminologies.html). */
export interface ValueSetFilter {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  property?: string;

  _property?: Element;

  op?:
    | "="
    | "is-a"
    | "descendent-of"
    | "is-not-a"
    | "regex"
    | "in"
    | "not-in"
    | "generalizes"
    | "exists";

  _op?: Element;

  value?: string;

  _value?: Element;
}